// libs/integration-tests/src/lib/model/container.enum.ts
export enum CONTAINER {
  POSTGRES = 'postgres',
  KEYCLOAK = 'keycloak',
  IAMKC_SVC = 'iamKcSvcContainer',
  WORKSPACE_SVC = 'workspaceSvcContainer',
  USER_PROFILE_SVC = 'userProfileSvcContainer',
  THEME_SVC = 'themeSvcContainer',
  TENANT_SVC = 'tenantSvcContainer',
  PRODUCT_STORE_SVC = 'productStoreSvcContainer',
  PERMISSION_SVC = 'permissionSvcContainer',
  SHELL_BFF = 'shellBff',
  SHELL_UI = 'shellUi',
  IMPORT_MANAGER = 'importManager',
}